Move XML-specific stuff to new XMLBlob class.
