+++ auslib/blobs/base.py
@@ -7,7 +7,6 @@ import jsonschema
 # To enable shared jsonschema validators
 import auslib.util.jsonschema_validators  # noqa
 import yaml
-from auslib.AUS import isSpecialURL
 from auslib.global_state import cache
 
 
@@ -158,20 +157,6 @@ class Blob(dict):
         if self.containsForbiddenDomain(product, whitelistedDomains):
             raise ValueError("Blob contains forbidden domain(s)")
 
-    def getResponseProducts(self):
-        """
-        :return: Usually returns None. If the Blob is a SuperBlob, it returns the list
-                of return products.
-        """
-        return None
-
-    def getResponseBlobs(self):
-        """
-        :return: Usually returns None. It the Blob is a systemaddons superblob, it returns the
-                 list of return blobs
-        """
-        return None
-
     def getSchema(self):
         def loadSchema():
             return yaml.safe_load(open(path.join(path.dirname(path.abspath(__file__)), "schemas", self.jsonschema)))
@@ -194,13 +179,33 @@ class Blob(dict):
         failing open)."""
         return False
 
-    def processSpecialForceHosts(self, url, specialForceHosts, force_arg):
-        if isSpecialURL(url, specialForceHosts):
-            if "?" in url:
-                url += "&force=" + force_arg.query_value
-            else:
-                url += "?force=" + force_arg.query_value
-        return url
+    def containsForbiddenDomain(self, product, whitelistedDomains):
+        raise NotImplementedError()
+
+    def getReferencedReleases(self):
+        """
+        :return: Returns set of names of partially referenced releases that the current
+        release references
+        """
+        return set()
+
+
+# TODO: getResponse* aren't technically XML-specific, but won't be used by other blobs
+# do they belong here, or on the base blob?
+class XMLBlob(Blob):
+    def getResponseProducts(self):
+        """
+        :return: Usually returns None. If the Blob is a SuperBlob, it returns the list
+                of return products.
+        """
+        return None
+
+    def getResponseBlobs(self):
+        """
+        :return: Usually returns None. It the Blob is a systemaddons superblob, it returns the
+                 list of return blobs
+        """
+        return None
 
     def getInnerHeaderXML(self, updateQuery, update_type, whitelistedDomains, specialForceHosts):
         """
@@ -217,9 +222,6 @@ class Blob(dict):
     def getInnerXML(self, updateQuery, update_type, whitelistedDomains, specialForceHosts):
         raise NotImplementedError()
 
-    def containsForbiddenDomain(self, product, whitelistedDomains):
-        raise NotImplementedError()
-
     def getHeaderXML(self):
         """
         :return: Returns the outer most header. Returns the outer most header
@@ -234,10 +236,3 @@ class Blob(dict):
         """
         footer = "</updates>"
         return footer
-
-    def getReferencedReleases(self):
-        """
-        :return: Returns set of names of partially referenced releases that the current
-        release references
-        """
-        return set()
